+++ ui-tests-msf/ui-tests-addv2/src/test/java/com/gooddata/qa/graphene/snowflake/SegmentDeleteTableForceLoadTest.java
@@ -154,9 +154,13 @@ public class SegmentDeleteTableForceLoadTest extends AbstractADDProcessTest {
         snowflakeUtils.createTable(DELETED_TABLE_CUSTOMERS, listDeleteColumnNormal);
         snowflakeUtils.createTable(DELETED_TABLE_CUSTOMERS_TIMESTAMP_CLIENTID, listDeleteColumnTimestampClientId);
         snowflakeUtils.createTable(DELETED_TABLE_CUSTOMERS_TIMESTAMP_DELETED, listDeleteColumnTimestampDeleted);
-        dataloadProcess = new ScheduleUtils(domainRestClient).createDataDistributionProcess(serviceProject, PROCESS_NAME,
-                dataSourceId, SEGMENT_ID, "att_lcm_default_data_product", "1");
-        domainProcessUtils = new ProcessUtils(domainRestClient, dataloadProcess);
+        try {
+            dataloadProcess = new ScheduleUtils(domainRestClient).createDataDistributionProcess(serviceProject, PROCESS_NAME,
+                    dataSourceId, SEGMENT_ID, "att_lcm_default_data_product", "1");
+            domainProcessUtils = new ProcessUtils(domainRestClient, dataloadProcess);
+        } catch (Exception e) {
+            throw new RuntimeException("Cannot create process" + e.getMessage());
+        }
         lcmBrickFlowBuilder.deleteMasterProject();
         lcmBrickFlowBuilder.runLcmFlow();
         lastSuccessful = LocalDateTime.now().withNano(0);
@@ -361,7 +365,9 @@ public class SegmentDeleteTableForceLoadTest extends AbstractADDProcessTest {
         if (testParams.getDeleteMode() == DeleteMode.DELETE_NEVER) {
             return;
         }
-        domainRestClient.getProcessService().removeProcess(dataloadProcess);
+        if (dataloadProcess != null) {
+            domainRestClient.getProcessService().removeProcess(dataloadProcess);
+        }
         lcmBrickFlowBuilder.destroy();
         dataSourceRestRequest.deleteDataSource(dataSourceId);
         snowflakeUtils.dropDatabaseIfExists(DATABASE_NAME);
